refactor, remove several libraries

+++ engine/src/consoleprint.cpp
@@ -11,13 +11,13 @@ void initConsole() {
 	AllocConsole();
 
 	HANDLE handle_out = GetStdHandle(STD_OUTPUT_HANDLE);
-	int hCrt = _open_osfhandle((long)handle_out, _O_TEXT);
+	int hCrt = _open_osfhandle((intptr_t)handle_out, _O_TEXT);
 	hf_out = _fdopen(hCrt, "w");
 	setvbuf(hf_out, NULL, _IONBF, 0);
 
 
 	HANDLE handle_in = GetStdHandle(STD_INPUT_HANDLE);
-	hCrt = _open_osfhandle((long)handle_in, _O_TEXT);
+	hCrt = _open_osfhandle((intptr_t)handle_in, _O_TEXT);
 	FILE* hf_in = _fdopen(hCrt, "r");
 	setvbuf(hf_in, NULL, _IONBF, 0);
 	*stdin = *hf_in;
